+++ src/index.ts
@@ -4,7 +4,12 @@ export { pluginCSS } from './utils/plugin-css';
 export { isCssImported } from './utils/plugin-css/loader';
 export { pluginDebug } from './utils/plugin-debug';
 
-export * as pluginTools from './utils/plugin-tools';
+export * as pluginTools from './utils/plugin-tools'; // All of the tools
+
+// The next lines are the plugin tools in separate namespaces
+export * as eventTools from './utils/plugin-tools/event-tools';
+export * as sectionTools from './utils/plugin-tools/section-tools';
+
 
 // Export types
 export type { RevealInstance } from './base/plugin-base';